Capitalized entries in .proto file resulted in the record name being overwritten.  For example given the protobuf file
message WithCaps {
  optional string nocaps = 1;
  optional string CAPS = 2;
}

if you compile and run without the patch, you get something like

1> rr(caps_pb).
[withcaps]
2> caps_pb:encode(#withcaps{caps="hello"}).
<<18,5,104,101,108,108,111>>
3> caps_pb:decode_withcaps(v(2)).
{"hello",undefined,undefined}

This fixes this in 2 ways.

First when determining the index, the not found case returns -1 instead of 0,
so it should be impossible to overwrite the record name, and you get a bad
argument exception.

Second the field is lowercased in the lookup table in the decode function.

Now the above works.

1> rr(caps_pb).
[withcaps]
2> caps_pb:encode(#withcaps{caps="hello"}).
<<18,5,104,101,108,108,111>>
3> caps_pb:decode_withcaps(v(2)).

Test to follow soon.