Fix default handling
Repeated fields are represented as a list in Erlang. Previously, if the
repeated field was empty, the decoder would leave the field as
'undefined' in the record which is counter intuitive and just
complicates handling of such fields. Rather than leaving the field
undefined its now set to an empty list as default.

Also the handling of defaults of optional fields is moved from the
encode to the decoder. Previously the encoder would always send the
default value of an optional field if it was undefined and the decoder
would leave the field undefined if it was not encoded in the message.
Now optional fields are only encoded if they differ from the default
value and the decoder will insert the default value if the field is not
encoded in the message.