Fix encoding of required fields with default value
The default value of a field must only be used for optional fields when
encoding. On the other hand, the default value is used for any field type
when decoding which effectively makes required fields with default value
optional for decoding. This seems to be in line with Google's protoc
compiler.