Ignore unknown fields when decoding.
The spec states that decoders should ignore fields that are
unrecognized. At a later time this could be refactored to stash
unrecognized fields in an additional field in the pbuf record.

NOTE: This is a reapplication of the commit from #11. There is one
known failure in the test suite, namely that binaries that are not
proper protobuffs messages will cause skip_next_field/1 to error with
a case clause error. This failure is caused by the call to
proper:check_specs() in module protobuffs_tests:proper_specs_test().