Ct tests will now run using proper or eqc
Ct does not use the modules compiled by rebar (during the eunit tests).
Ct will try to compile the *_SUITE test modules. It does this by trying to
find an "Emakefile" in the same directory as the suite. If there isn't one,
it just compiles everything using default options. This meant that the
tests rebar does to set to the 'EQC' and 'PROPER' macros is never run or
found. So, the quickcheck tests that ct defines where never run.

The change adds a script to generate the Emakefile before the ct tests are
run. It uses similar logic as rebar to accomplish the task. Note how tests
are failing (as of this commit); at least now they are running.