Added deliminated encode and decode functionality.
In protobuffs version 2 for java, there's a set of deliminated io
functions. This makes it relatively convienent for streamed messages. The
encode functions have been extended to accept lists of records, and there's
a deliminated_decode_'<record>' added which returns a list of records
(and any left over from the passed in binary).